Break out agent commands
